Common: Rename types.plugin_name to types.plugin_types
