+++ monkey/common/types/__init__.py
@@ -2,5 +2,5 @@ from .concurrency import Lock, Event
 from .serialization import JSONSerializable
 from .ids import AgentID, HardwareID, MachineID
 from .networking import NetworkService, NetworkPort, PortStatus, SocketAddress
-from .plugin_name import PluginName
-from .plugin_name import PluginVersion
+from .plugin_types import PluginName
+from .plugin_types import PluginVersion
